requirements.txt
absl-py
astunparse
audioread
certifi
cffi
charset-normalizer
contourpy
cycler
decorator
flatbuffers
fonttools
gast
google-pasta
grpcio
h5py
idna
joblib
keras
kiwisolver
lazy_loader
libclang
librosa
llvmlite
Markdown
markdown-it-py
MarkupSafe
matplotlib
mdurl
ml-dtypes
msgpack
namex
numba
numpy
opt_einsum
optree
packaging
pandas
pillow
platformdirs
pooch
protobuf
pycparser
Pygments
pyparsing
python-dateutil
pytz
requests
rich
scikit-learn
scipy
six
soundfile
soxr
tensorboard
tensorboard-data-server
tensorflow
tensorflow-io-gcs-filesystem
tensorflow-macos
tensorflow-metal
termcolor
threadpoolctl
typing_extensions
tzdata
urllib3
webrtcvad
Werkzeug
wrapt
Flask-Migrate
